ai-log-sync/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-log-sync"
version = "0.1.0"
description = "Aggregate AI conversation logs from multiple platforms"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
ai-log-sync = "ai_log_sync.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_log_sync"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
